Undo changes to diagnostic_interface (1)
